pyproject.toml
[project]
name = "flow-simulation-pinn"
version = "0.1.0"
description = "Two-phase flow simulation in porous media using Physics-Informed Neural Networks"
readme = "README.md"
requires-python = "==3.12.11"
dependencies = [
    "comet-ml==3.55.0",
    "flax==0.12.2",
    "grain==0.2.15",
    "hydra-core==1.3.2",
    "jax==0.8.2",
    "matplotlib==3.10.8",
    "numpy==2.4.0",
    "optax==0.2.6",
    "pandas==2.3.3",
    "scipy==1.16.3",
    "tqdm==4.67.1",
]

[dependency-groups]
dev = [
    "pre-commit==4.5.1",
    "ruff==0.14.9",
    "ty==0.0.3",
]

[tool.uv]
package = true

[tool.setuptools]
packages = ["src"]

[tool.ruff.format]
quote-style = "single"

[tool.ruff]
line-length = 100
